Fix for shift issue (SMB jump noise)
Fix for issue where a "runaway" pitch shift would 
cause an incorrect note, or no note, to be 
sounded. This happened on the SMB jump noise. The 
issue was caused by the pitch shift overflowing, 
thereby incorrectly setting the prescaler.

The fix involves always resetting the pitch shift 
before setting the prescaler value.