Make MIDI receive processing more sophisticated
MIDI was sometimes receiving a spurious NoteOn 
message (always a very low note, and received 
after holding down an arpeggio for a while -- 
amount of hold time varied).

The problem was partially addressed by verifying 
the MIDI packet CableNumber, but this just 
increased the average time before the problem 
occurred, implying that the CableNumber was random
in the spurious message.

By also validating the MIDI channel number 
(currently set at 0 via a #define), the problem 
seems to be solved (although it's possible that, 
just as checking cableNumber reduced the frequency
of occurence, so too does checking the channel 
reduce the frequency by a factor of 16).