Arp click removal and general code efficiency
Get rid of the annoying audible "click" during 
fast arpeggiation. This was achieved by making the
arp interrupt non-blocking. I tried that before 
without success, which turned out to be because 
the arp clock took too long. The solution was to
change the arp interrupt speed. An even better 
solution would be to move the arp interrupt to the
16-bit Timer3.

Speaking of which, more efficiency could be found 
by setting the speed more intelligently on Timer0.