Fix for bug caused by button mashing over chord
If a chord was held and another button 
(or buttons) was pressed very rapidly, things 
would very quickly go nuts: weird sounds, global
pitch shift, etc. All the symptoms of an overflow.

If the NOBLOCK statement was removed from the arp 
interrupt, the problem would disappear, but the 
clicking during chords would return (see 332521b74969d1c9b80a99dba9c3a80a504cec11).

The only thing that could cause an overflow seemed 
to be the chord length changing, which is only 
happening in one place, so I wrapped that in a 
cli()/sei() pair, and it appears to have done the 
trick.

Also made chord_length volatile; not sure if that 
is actually necessary, but it doesn't seem to be 
doing any harm.